refactor: wrap logo and update tag
