+++ packages/frontend/src/app/api/auth/change-password/page.tsx
@@ -1,7 +1,6 @@
 import { Metadata } from 'next';
 import { ChangePasswordForm } from '@/components/change-password-form';
-import Image from 'next/image';
-import Logo from '@/../public/logo.svg';
+import { Logo } from '@/components/logo';
 import Link from 'next/link';
 
 export const metadata: Metadata = {
@@ -14,20 +13,15 @@ export default function ChangePassword() {
     <div className="change-password-container container my-10 flex flex-col items-center justify-center">
       <div className="change-password-card h-auto w-full rounded-lg border bg-white py-8 md:max-w-md">
         <div className="change-password-content w-ful mx-auto flex flex-col justify-center space-y-6 sm:w-[350px]">
-          <div className="change-password-header flex flex-col space-y-5 text-center">
-            <Image
-              src={Logo}
-              alt="logo"
-              className="mx-auto h-10 w-24 object-contain"
-            />
-
+          <header className="change-password-header flex flex-col space-y-5 text-center">
+            <Logo className="mx-auto h-10 w-24 object-contain" />
             <h1 className="font-subheading text-2xl font-semibold tracking-tight">
               Change Password
             </h1>
             <p className="text-sm text-muted-foreground">
               Enter your new password below
             </p>
-          </div>
+          </header>
           <ChangePasswordForm />
           <div className="reset-actions flex flex-col gap-4 text-center text-sm text-muted-foreground">
             <span className="px-5">
