+++ packages/frontend/src/app/api/auth/register/page.tsx
@@ -1,8 +1,7 @@
 import { Metadata } from 'next';
 import Link from 'next/link';
 import { UserAuthForm } from '@/components/user-auth-form';
-import Image from 'next/image';
-import Logo from '@/../public/logo.svg';
+import { Logo } from '@/components/logo';
 
 export const metadata: Metadata = {
   title: 'Register your account',
@@ -14,19 +13,15 @@ export default function Register() {
     <div className="register-container container my-10 flex flex-col items-center justify-center">
       <div className="register-card h-auto w-full rounded-lg border bg-white py-8 md:max-w-md">
         <div className="register-content w-ful mx-auto flex flex-col justify-center space-y-6 sm:w-[350px]">
-          <div className="register-header flex flex-col space-y-5 text-center">
-            <Image
-              src={Logo}
-              alt="logo"
-              className="mx-auto h-10 w-24 object-contain"
-            />
+          <header className="register-header flex flex-col space-y-5 text-center">
+            <Logo className="mx-auto h-10 w-24 object-contain" />
             <h1 className="font-subheading text-2xl font-semibold tracking-tight">
               Create an Account
             </h1>
             <p className="px-12 text-sm text-muted-foreground">
               Enter your email and password below to create your account
             </p>
-          </div>
+          </header>
           <UserAuthForm />
           <div className="login-actions flex flex-col gap-4 text-center text-sm text-muted-foreground">
             <span className="px-5">
